adjust change detection in rest updates

+++ server/grails-app/services/gokbg3/RestMappingService.groovy
@@ -296,8 +296,9 @@ class RestMappingService {
    */
 
   @Transactional
-  def updateObject(obj, jsonMap, reqBody) {
+  public Boolean updateObject(obj, jsonMap, reqBody) {
     log.debug("Update object ${obj} - ${reqBody}")
+    Boolean changed = false
     PersistentEntity pent = grailsApplication.mappingContext.getPersistentEntity(obj.class.name)
 
     def toIgnore = defaultIgnore + (jsonMap?.ignore ?: [])
@@ -311,7 +312,7 @@ class RestMappingService {
 
         if (p instanceof Association) {
           if (p instanceof ManyToOne || p instanceof OneToOne) {
-            updateAssoc(obj, p.name, newVal)
+            changed |= updateAssoc(obj, p.name, newVal)
           }
           else {
             // Add to collection
@@ -339,15 +340,19 @@ class RestMappingService {
         }
       }
     }
-    if(obj.validate()) {
-      obj.save()
-    } else {
-      obj
+    if (obj.isDirty()) {
+      changed = true
     }
+
+    if (obj.validate()) {
+      obj.save()
+    }
+
+    return changed
   }
 
-  @Transactional
-  def updateAssoc(obj, prop, val, def cat = null) {
+  private boolean updateAssoc(obj, prop, val, def cat = null) {
+    boolean changed = false
     log.debug("Update association $obj - $prop: $val")
     def ptype = grailsApplication.mappingContext.getPersistentEntity(obj.class.name).getPropertyByName(prop).type
 
@@ -377,10 +382,11 @@ class RestMappingService {
             if (rdv) {
               if (rdv in cat.values) {
                 if (catName == 'KBComponent.Status') {
-                  updateStatus(obj, rdv.value)
+                  changed = updateStatus(obj, rdv.value)
                 }
-                else {
+                else if (rdv != obj[prop]) {
                   obj[prop] = rdv
+                  changed = true
                 }
               }
               else {
@@ -415,10 +421,11 @@ class RestMappingService {
               if (rdv) {
                 if (rdv in cat.values) {
                   if (catName == 'KBComponent.Status') {
-                    updateStatus(obj, rdv.value)
+                    changed = updateStatus(obj, rdv.value)
                   }
-                  else {
+                  else if (rdv != obj[prop]) {
                     obj[prop] = rdv
+                    changed = true
                   }
                 }
                 else {
@@ -464,10 +471,11 @@ class RestMappingService {
               }
               else {
                 if (catName == 'KBComponent.Status') {
-                  updateStatus(obj, rdv.value)
+                  changed = updateStatus(obj, rdv.value)
                 }
-                else {
+                else if (rdv != obj[prop]) {
                   obj[prop] = rdv
+                  changed = true
                 }
               }
             }
@@ -492,10 +500,11 @@ class RestMappingService {
             }
             else {
               if (catName == 'KBComponent.Status') {
-                updateStatus(obj, rdv.value)
+                changed = updateStatus(obj, rdv.value)
               }
-              else {
+              else if (rdv != obj[prop]) {
                 obj[prop] = rdv
+                changed = true
               }
             }
           }
@@ -516,7 +525,9 @@ class RestMappingService {
         }
 
         if (linkObj) {
-          obj[prop] = linkObj
+          if (linkObj != obj[prop]) {
+            obj[prop] = linkObj
+          }
         }
         else {
           obj.errors.reject(
@@ -531,12 +542,13 @@ class RestMappingService {
         }
       }
     }
-    else {
+    else if (obj[prop] != null) {
       log.debug("Set value to null")
+      changed = true
       obj[prop] = null
     }
 
-    obj
+    changed
   }
 
   @Transactional
@@ -669,21 +681,34 @@ class RestMappingService {
     result
   }
 
-  @Transactional
-  public def updateStatus(obj, val) {
-    if (val == 'Deleted') {
-      obj.deleteSoft()
+  public boolean updateStatus(obj, val) {
+    boolean changed = false
 
-      componentUpdateService.closeConnectedReviews(obj)
+    if (val == 'Deleted') {
+      if (obj.status.value != 'Deleted') {
+        changed = true
+        obj.deleteSoft()
+
+        componentUpdateService.closeConnectedReviews(obj)
+      }
     }
     else if (val == 'Retired') {
-      obj.retire()
+      if (obj.status.value != 'Retired' ) {
+        changed = true
+        obj.retire()
+      }
     }
     else if (val == 'Current') {
-      obj.setActive()
+      if (obj.status.value != 'Current') {
+        changed = true
+        obj.setActive()
+      }
     }
     else if (val == 'Expected') {
-      obj.setExpected()
+      if (obj.status.value != 'Expected') {
+        changed = true
+        obj.setExpected()
+      }
     }
     else {
       obj.errors.reject(
@@ -697,7 +722,7 @@ class RestMappingService {
       )
     }
 
-    obj
+    changed
   }
 
 
@@ -765,7 +790,6 @@ class RestMappingService {
   public def updateVariantNames(obj, vals, boolean remove = true) {
     log.debug("Update Variants ${vals} ..")
     def result = [changed: false, errors: []]
-    def changed = false
     def remaining = []
     def notFound = []
     def toRemove = []
@@ -788,7 +812,7 @@ class RestMappingService {
 
                 if (newVariant) {
                   log.debug("Added variant ${newVariant}")
-                  changed = true
+                  result.changed = true
                   remaining << newVariant
                 }
                 else {
@@ -838,7 +862,7 @@ class RestMappingService {
                 log.debug("Ensured variant: ${newVariant}")
 
                 if (newVariant) {
-                  changed = true
+                  result.changed = true
 
                   if (it.locale) {
                     newVariant = updateAssoc(newVariant, 'locale', it.locale, RefdataCategory.findByDesc(KBComponent.RD_LANGUAGE))
@@ -882,7 +906,7 @@ class RestMappingService {
             obj.variantNames.each { vn ->
               if (!remaining.contains(vn)) {
                 toRemove.add(vn.id)
-                changed = true
+                result.changed = true
               }
             }
 
@@ -901,7 +925,7 @@ class RestMappingService {
           }
         }
 
-        if (changed) {
+        if (result.changed) {
           obj.lastSeen = System.currentTimeMillis()
         }
       }
@@ -965,6 +989,7 @@ class RestMappingService {
               ComponentPrice cp_to_delete = ComponentPrice.get(ep)
               obj.removeFromPrices(cp_to_delete)
               cp_to_delete.delete()
+              result.changed = true
             }
           }
         }
@@ -1150,7 +1175,7 @@ class RestMappingService {
     result
   }
 
-  public def updateLongField(obj, prop, val) {
+  public void updateLongField(obj, prop, val) {
     log.debug("Set simple prop ${prop} = ${val} (as Long)")
 
     try {
@@ -1167,10 +1192,9 @@ class RestMappingService {
           'typeMismatch.java.lang.Long'
       )
     }
-    obj
   }
 
-  public def updateDateField(obj, prop, val) {
+  public void updateDateField(obj, prop, val) {
     if (val == null || !val.trim()) {
       obj[prop] = null
     }
@@ -1193,7 +1217,6 @@ class RestMappingService {
       }
       log.debug("Set simple prop ${prop} = ${val} (as date ${dateObj}))")
     }
-    obj
   }
 
   /**
